Changed input method
Button/movement actions may now be entered in any order, e.g. the A/B button may be pressed before or after the direction action (arrow key or joystick).

I moved P1_a and P1_b to the A and S keys to avoid "ghosting" (some combinations of 3 keys do not work on my keyboard). For P2 there are still some keyboard combinations that don't work...