Cleaned up input events
Alt-Enter works everywhere (toggles full-screen mode).

The code for testing for a key or button press to pass to the next screen is now in one place (PreloadScript01).

In general, DON'T call set_process_input(false), so that Alt-Enter events are handles correctly.

In _input() functions, use the event passed to the function rather than polling Input.